Добавил Specifications Обновил сервисы для проверки доступности информации пользователю

+++ WorkyOne.AppServices/Services/Schedule/Common/TemplateService.cs
@@ -1,15 +1,22 @@
-﻿using System.Net.WebSockets;
-using System.Text;
+﻿using System.Text;
 using AutoMapper;
 using WorkyOne.AppServices.Interfaces.Repositories.Schedule.Common;
+using WorkyOne.AppServices.Interfaces.Repositories.Schedule.Shifts;
 using WorkyOne.AppServices.Interfaces.Services.Schedule.Common;
+using WorkyOne.AppServices.Interfaces.Services.Schedule.Users;
 using WorkyOne.AppServices.Interfaces.Utilities;
 using WorkyOne.Contracts.DTOs.Schedule.Common;
 using WorkyOne.Contracts.Services.Common;
 using WorkyOne.Contracts.Services.CreateModels.Schedule.Common;
+using WorkyOne.Contracts.Services.GetRequests.Common;
 using WorkyOne.Domain.Entities.Schedule.Common;
+using WorkyOne.Domain.Entities.Schedule.Shifts;
 using WorkyOne.Domain.Requests.Common;
 using WorkyOne.Domain.Requests.Schedule.Common;
+using WorkyOne.Domain.Specifications.AccesFilters.Schedule.Common;
+using WorkyOne.Domain.Specifications.AccesFilters.Schedule.Shifts;
+using WorkyOne.Domain.Specifications.Base;
+using WorkyOne.Domain.Specifications.Common;
 
 namespace WorkyOne.AppServices.Services.Schedule.Common
 {
@@ -21,15 +28,25 @@ namespace WorkyOne.AppServices.Services.Schedule.Common
         private readonly ITemplatesRepository _templatesRepo;
         private readonly ISchedulesRepository _schedulesRepo;
         private readonly IShiftSequencesRepository _shiftsSequencesRepo;
+        private readonly ITemplatedShiftsRepository _templatedShiftsRepo;
+
         private readonly IMapper _mapper;
         private readonly IEntityUpdateUtility _updateUtility;
+        private readonly IUserAccessInfoProvider _accessInfoProvider;
+
+        private TemplateAccessFilter _templateFilter;
+        private ScheduleAccessFilter _scheduleFilter;
+        private ShiftSequenceAccessFilter _sequenceFilter;
+        private TemplatedShiftAccessFilter _shiftFilter;
 
         public TemplateService(
             ITemplatesRepository templatesRepo,
             IShiftSequencesRepository shiftsSequencesRepo,
             IMapper mapper,
             IEntityUpdateUtility updateUtility,
-            ISchedulesRepository schedulesRepo
+            ISchedulesRepository schedulesRepo,
+            IUserAccessInfoProvider accessInfoProvider,
+            ITemplatedShiftsRepository templatedShiftsRepo
         )
         {
             _templatesRepo = templatesRepo;
@@ -37,6 +54,10 @@ namespace WorkyOne.AppServices.Services.Schedule.Common
             _mapper = mapper;
             _updateUtility = updateUtility;
             _schedulesRepo = schedulesRepo;
+            _accessInfoProvider = accessInfoProvider;
+
+            InitFiltersAsync().Wait();
+            _templatedShiftsRepo = templatedShiftsRepo;
         }
 
         public async Task<ServiceResult> CreateAsync(
@@ -44,10 +65,11 @@ namespace WorkyOne.AppServices.Services.Schedule.Common
             CancellationToken cancellation = default
         )
         {
-            var schedule = await _schedulesRepo.GetAsync(
-                new ScheduleRequest(model.ScheduleId, true),
-                cancellation
-            );
+            var filter = new EntityIdFilter<ScheduleEntity>(model.ScheduleId).And(_scheduleFilter);
+
+            var request = new ScheduleRequest(filter) { IncludeTemplate = true };
+
+            var schedule = await _schedulesRepo.GetAsync(request, cancellation);
 
             if (schedule == null)
             {
@@ -90,12 +112,17 @@ namespace WorkyOne.AppServices.Services.Schedule.Common
             CancellationToken cancellation = default
         )
         {
-            var result = await _templatesRepo.DeleteAsync(id, cancellation);
+            var filter = new EntityIdFilter<TemplateEntity>(id).And(_templateFilter);
 
-            if (cancellation.IsCancellationRequested)
+            var request = new EntityRequest<TemplateEntity>(filter);
+            var deleted = await _templatesRepo.GetAsync(request, cancellation);
+
+            if (deleted == null)
             {
-                return ServiceResult.CancellationRequested();
+                return ServiceResult.Error($"Сущность (ID: {id}) не найдена, либо доступ запрещён");
             }
+
+            var result = _templatesRepo.Delete(deleted);
 
             if (result.IsSuccess)
             {
@@ -113,9 +140,12 @@ namespace WorkyOne.AppServices.Services.Schedule.Common
             CancellationToken cancellation = default
         )
         {
-            var request = new EntityRequest<TemplateEntity>(id);
-            var entity = await _templatesRepo.GetAsync(request, cancellation);
+            var filter = new EntityIdFilter<TemplateEntity>(id).And(_templateFilter);
 
+            var entity = await _templatesRepo.GetAsync(
+                new EntityRequest<TemplateEntity>(filter),
+                cancellation
+            );
             if (entity == null)
             {
                 return null;
@@ -130,12 +160,14 @@ namespace WorkyOne.AppServices.Services.Schedule.Common
             CancellationToken cancellation = default
         )
         {
-            var request = new EntityRequest<TemplateEntity>()
-            {
-                Predicate = (x) => x.ScheduleId == scheduleId
-            };
+            var filter = new Specification<TemplateEntity>(x => x.ScheduleId == scheduleId).And(
+                _templateFilter
+            );
 
-            var entity = await _templatesRepo.GetAsync(request, cancellation);
+            var entity = await _templatesRepo.GetAsync(
+                new EntityRequest<TemplateEntity>(filter),
+                cancellation
+            );
 
             if (entity == null)
             {
@@ -146,31 +178,63 @@ namespace WorkyOne.AppServices.Services.Schedule.Common
             return dto;
         }
 
+        public async Task<List<ShiftSequenceDto>> GetSequencesAsync(
+            string templateId,
+            PaginatedRequest request,
+            CancellationToken cancellation = default
+        )
+        {
+            var filter = new Specification<ShiftSequenceEntity>(x =>
+                x.TemplateId == templateId
+            ).And(_sequenceFilter);
+
+            var entities = await _shiftsSequencesRepo.GetManyAsync(
+                new PaginatedRequest<ShiftSequenceEntity>(
+                    filter,
+                    request.PageIndex,
+                    request.Amount
+                ),
+                cancellation
+            );
+
+            var dtos = _mapper.Map<List<ShiftSequenceDto>>(entities);
+
+            dtos = dtos.OrderBy(x => x.Position).ToList();
+
+            return dtos;
+        }
+
         public async Task<ServiceResult> UpdateAsync(
             TemplateDto dto,
             CancellationToken cancellation = default
         )
         {
-            var source = _mapper.Map<TemplateEntity>(dto);
+            var filter = new EntityIdFilter<TemplateEntity>(dto.Id).And(_templateFilter);
 
-            var request = new EntityRequest<TemplateEntity>(source.Id);
-            var target = await _templatesRepo.GetAsync(request, cancellation);
+            var target = await _templatesRepo.GetAsync(
+                new EntityRequest<TemplateEntity>(filter),
+                cancellation
+            );
 
             if (target == null)
             {
-                return ServiceResult.Error($"Обновляемый шаблон (ID: {source.Id}) не найден");
+                return ServiceResult.Error(
+                    $"Обновляемый шаблон (ID: {dto.Id}) не найден либо доступ запрещён"
+                );
             }
 
-            if (cancellation.IsCancellationRequested)
-            {
-                return ServiceResult.CancellationRequested();
-            }
+            var source = _mapper.Map<TemplateEntity>(dto);
 
             _updateUtility.Update(target, source);
 
             _templatesRepo.Update(target);
 
             await _templatesRepo.SaveChangesAsync(cancellation);
+
+            if (cancellation.IsCancellationRequested)
+            {
+                return ServiceResult.CancellationRequested();
+            }
 
             return ServiceResult.Ok($"Шаблон (ID: {target.Id}) был успешно обновлён");
         }
@@ -180,22 +244,63 @@ namespace WorkyOne.AppServices.Services.Schedule.Common
             CancellationToken cancellation = default
         )
         {
-            var request = new EntityRequest<TemplateEntity>(model.TemplateId);
-            var template = await _templatesRepo.GetAsync(request, cancellation);
+            var filter = new EntityIdFilter<TemplateEntity>(model.TemplateId).And(_templateFilter);
 
-            if (template == null)
-            {
-                return ServiceResult.Error($"Не найден шаблон(ID: {model.TemplateId})");
-            }
-
-            await _shiftsSequencesRepo.DeleteByConditionAsync(
-                (x) => x.TemplateId == template.Id,
+            var template = await _templatesRepo.GetAsync(
+                new EntityRequest<TemplateEntity>(filter),
                 cancellation
             );
 
-            var sequence = _mapper.Map<List<ShiftSequenceEntity>>(model.Sequences);
-            sequence.ForEach(x => x.Template = template);
+            if (template == null)
+            {
+                return ServiceResult.Error(
+                    $"Не найден шаблон(ID: {model.TemplateId}), либо доступ запрещён"
+                );
+            }
 
+            model.Sequences = model.Sequences.OrderBy(x => x.Position).ToList();
+
+            if (!CheckSequence(model.Sequences))
+            {
+                return ServiceResult.Error(
+                    $"Значение Position должно быть уникально для каждой сущности"
+                );
+            }
+
+            if (!model.Sequences.Any())
+            {
+                await _shiftsSequencesRepo.DeleteByConditionAsync(
+                    new Specification<ShiftSequenceEntity>(x => x.TemplateId == template.Id),
+                    cancellation
+                );
+
+                return ServiceResult.Ok(
+                    $"Последовательность для шаблона (ID: {template.Id}) была успешно очищена!"
+                );
+            }
+
+            var sequence = _mapper.Map<List<ShiftSequenceEntity>>(model.Sequences);
+            var shifts = await GetTemplatedShiftAsync(template.Id, cancellation);
+
+            foreach (var item in sequence)
+            {
+                item.Template = template;
+
+                var shift = shifts.First(x => x.Id == item.ShiftId);
+                if (shift == null)
+                {
+                    return ServiceResult.Error(
+                        $"Не удалось найти \"шаблонную\" смену (ID: {item.ShiftId})"
+                    );
+                }
+                item.Shift = shift;
+            }
+            sequence = sequence.OrderBy(x => x.Position).ToList();
+
+            await _shiftsSequencesRepo.DeleteByConditionAsync(
+                new Specification<ShiftSequenceEntity>(x => x.TemplateId == template.Id),
+                cancellation
+            );
             var result = await _shiftsSequencesRepo.CreateManyAsync(sequence, cancellation);
 
             if (cancellation.IsCancellationRequested)
@@ -217,6 +322,49 @@ namespace WorkyOne.AppServices.Services.Schedule.Common
             {
                 return ServiceResult.FromRepositoryResult(result);
             }
+        }
+
+        private async Task InitFiltersAsync()
+        {
+            if (_sequenceFilter == null)
+            {
+                var accessInfo = await _accessInfoProvider.GetCurrentAsync();
+
+                _sequenceFilter = new ShiftSequenceAccessFilter(accessInfo);
+                _templateFilter = new TemplateAccessFilter(accessInfo);
+                _scheduleFilter = new ScheduleAccessFilter(accessInfo);
+                _shiftFilter = new TemplatedShiftAccessFilter(accessInfo);
+            }
+        }
+
+        private bool CheckSequence(IEnumerable<ShiftSequenceDto> sequence)
+        {
+            var i = 0;
+            foreach (var item in sequence)
+            {
+                if (item.Position == i)
+                {
+                    return false;
+                }
+
+                item.Position = ++i;
+            }
+
+            return true;
+        }
+
+        private Task<List<TemplatedShiftEntity>> GetTemplatedShiftAsync(
+            string templateId,
+            CancellationToken cancellation = default
+        )
+        {
+            var filter = new Specification<TemplatedShiftEntity>(x =>
+                x.TemplateId == templateId
+            ).And(_shiftFilter);
+
+            return _templatedShiftsRepo.GetManyAsync(
+                new PaginatedRequest<TemplatedShiftEntity>(filter, 1, 100)
+            );
         }
     }
 }
